1.2.1 to master (#24)
* add `show_only_login` column, and add new route for login user (#9)

* Send reservation email (#10)

* update JWT token property

* update ESLint option

* add place booker email

* add equip booker email

* [feat] add community center

* [fix] remove unnecessary change

* [feat] add error logic for inappropriate reservation

* implt user reservation delete API (#13)

* implt introduction views and views add (#14)

* update introduction sorting rule

* Popo setting json (#15)

* setup setting module

* implement popo setting API and s3 setup

* Place reservation accept all (#16)

* implt multiple place reservation accept

* refactor reservation possible check

* implt equip reservation overlap check

* [feat] Implement community center auto accept logic

* [feat] add max time column in place table and implement max time check logic

* Search api (#18)

* implt user search API

* fix query mistake

* user search final implt

* fix notation and some logic

* [feat] add max minutes column to equipment and implement time check logic

* fix order of if statement

* add link columns (#20)

* [feat] Implement logic to allow reservations at the same time as rejected reservations (#21)

* Fix/max minutes (#22)

* update dto class name

* place/equip default max_minutes to 24 * 60

* polish service `update` logic

* refactor reservation with max_minutes checking

* simplify overlap condition

* implt user password reset API

* opening hours (#23)

* update package-lock

* add opening_hours column

* improve code

* add opening_hours column

* add comment

---------

+++ src/popo/reservation/equip/reserve.equip.controller.ts
@@ -8,6 +8,7 @@ import {
   Post,
   Query,
   Req,
+  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
 import { ApiQuery, ApiTags } from '@nestjs/swagger';
@@ -51,6 +52,12 @@ export class ReserveEquipController {
         existEquips,
         new_reservation,
       ),
+    );
+
+    // send e-mail to booker
+    await this.mailService.sendEquipReserveCreateMailToBooker(
+      user.email,
+      new_reservation,
     );
 
     return new_reservation;
@@ -127,8 +134,19 @@ export class ReserveEquipController {
 
   @Delete(':uuid')
   @UseGuards(JwtAuthGuard)
-  delete(@Param('uuid') uuid: string) {
-    return this.reserveEquipService.remove(uuid);
+  async delete(@Param('uuid') uuid: string, @Req() req) {
+    const reservation = await this.reserveEquipService.findOne(uuid);
+    const user = req.user;
+
+    if (user.userType == UserType.admin || user.userType == UserType.staff) {
+      return this.reserveEquipService.remove(uuid);
+    } else {
+      if (reservation.booker_id == user.uuid) {
+        return this.reserveEquipService.remove(uuid);
+      } else {
+        throw new UnauthorizedException('Unauthorized delete action');
+      }
+    }
   }
 
   /**
