+++ src/popo/reservation/place/reserve.place.controller.ts
@@ -8,13 +8,16 @@ import {
   Post,
   Query,
   Req,
+  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
 import { ApiQuery, ApiTags } from '@nestjs/swagger';
 import { Request } from 'express';
-
 import { ReservePlaceService } from './reserve.place.service';
-import { CreateReservePlaceDto } from './reserve.place.dto';
+import {
+  AcceptPlaceReservationListDto,
+  CreateReservePlaceDto,
+} from './reserve.place.dto';
 import { MailService } from '../../../mail/mail.service';
 import { ReservationStatus } from '../reservation.meta';
 import { UserType } from '../../user/user.meta';
@@ -36,15 +39,22 @@ export class ReservePlaceController {
   @UseGuards(JwtAuthGuard)
   async createWithNameAndId(@Req() req, @Body() dto: CreateReservePlaceDto) {
     const user: any = req.user;
+    const existPlace = await this.placeService.findOneOrFail(dto.place_id);
 
-    const saveDto = Object.assign(dto, { booker_id: user.uuid });
-    const new_reservation = await this.reservePlaceService.save(saveDto);
+    const new_reservation = await this.reservePlaceService.save(
+      Object.assign(dto, { booker_id: user.uuid }),
+    );
 
-    const existPlace = await this.placeService.findOne(dto.place_id);
-
-    // Send e-mail to staff.
-    this.mailService.sendPlaceReserveCreateMailToStaff(
+    // Send e-mail to staff
+    await this.mailService.sendPlaceReserveCreateMailToStaff(
       existPlace.staff_email,
+      existPlace,
+      new_reservation,
+    );
+
+    // Send e-mail to booker
+    await this.mailService.sendPlaceReserveCreateMailToBooker(
+      user.email,
       existPlace,
       new_reservation,
     );
@@ -144,6 +154,33 @@ export class ReservePlaceController {
     return this.reservePlaceService.findAllByPlaceName(placeName);
   }
 
+  @Patch('all/status/accept')
+  @Roles(UserType.admin, UserType.association, UserType.staff)
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  async acceptAllStatus(
+    @Body() body: AcceptPlaceReservationListDto,
+    @Query('sendEmail') sendEmail?: string,
+  ) {
+    for (const reservation_uuid of body.uuid_list) {
+      const response = await this.reservePlaceService.updateStatus(
+        reservation_uuid,
+        ReservationStatus.accept,
+      );
+
+      if (sendEmail === 'true') {
+        // Send e-mail to client.
+        const skipList = [UserType.admin, UserType.association, UserType.club];
+        if (!skipList.includes(response.userType)) {
+          await this.mailService.sendReservationPatchMail(
+            response.email,
+            response.title,
+            ReservationStatus[status],
+          );
+        }
+      }
+    }
+  }
+
   @Patch(':uuid/status/:status')
   @Roles(UserType.admin, UserType.association, UserType.staff)
   @UseGuards(JwtAuthGuard, RolesGuard)
@@ -171,9 +208,19 @@ export class ReservePlaceController {
   }
 
   @Delete(':uuid')
-  @Roles(UserType.admin, UserType.association, UserType.staff)
-  @UseGuards(JwtAuthGuard, RolesGuard)
-  delete(@Param('uuid') uuid: string) {
-    return this.reservePlaceService.remove(uuid);
+  @UseGuards(JwtAuthGuard)
+  async delete(@Param('uuid') uuid: string, @Req() req) {
+    const reservation = await this.reservePlaceService.findOne(uuid);
+    const user = req.user;
+
+    if (user.userType == UserType.admin || user.userType == UserType.staff) {
+      return this.reservePlaceService.remove(uuid);
+    } else {
+      if (reservation.booker_id == user.uuid) {
+        return this.reservePlaceService.remove(uuid);
+      } else {
+        throw new UnauthorizedException('Unauthorized delete action');
+      }
+    }
   }
 }
